+++ __test__/fixtures/exports-versioned-types-stub/old/outdated.d.ts
@@ -1,0 +1,1 @@
+export declare const upgradeYourTypeScript: 'this is a stub for TypeScript below 5.4';